+++ install/src/main/java/org/jetbrains/bsp/bazel/install/cli/CliOptions.kt
@@ -19,9 +19,14 @@ data class ProjectViewCliOptions internal constructor(
         val importDepth: Int?,
 )
 
+data class BloopCliOptions internal constructor(
+        val useBloop: Boolean
+)
+
 data class CliOptions internal constructor(
         val helpCliOptions: HelpCliOptions,
         val workspaceRootDir: Path,
         val projectViewFilePath: Path?,
         val projectViewCliOptions: ProjectViewCliOptions?,
+        val bloopCliOptions: BloopCliOptions,
 )