+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/EmptyLanguagePlugin.java
@@ -2,6 +2,7 @@ package org.jetbrains.bsp.bazel.server.sync.languages;
 
 import ch.epfl.scala.bsp4j.BuildTarget;
 import io.vavr.control.Option;
+import java.nio.file.Path;
 import org.jetbrains.bsp.bazel.info.BspTargetInfo;
 
 public class EmptyLanguagePlugin extends LanguagePlugin<LanguageData> {
@@ -12,4 +13,9 @@ public class EmptyLanguagePlugin extends LanguagePlugin<LanguageData> {
 
   @Override
   protected void applyModuleData(LanguageData moduleData, BuildTarget buildTarget) {}
+
+  @Override
+  public Option<Path> calculateSourceRoot(Path source) {
+    return Option.none();
+  }
 }