+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/LanguagePlugin.java
@@ -6,6 +6,7 @@ import io.vavr.collection.Seq;
 import io.vavr.collection.Set;
 import io.vavr.control.Option;
 import java.net.URI;
+import java.nio.file.Path;
 import org.jetbrains.bsp.bazel.info.BspTargetInfo.TargetInfo;
 import org.jetbrains.bsp.bazel.server.sync.dependencytree.DependencyTree;
 
@@ -25,4 +26,6 @@ public abstract class LanguagePlugin<T extends LanguageData> {
   }
 
   protected abstract void applyModuleData(T moduleData, BuildTarget buildTarget);
+
+  public abstract Option<Path> calculateSourceRoot(Path source);
 }