+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/languages/scala/ScalaLanguagePlugin.java
@@ -16,11 +16,13 @@ import io.vavr.collection.Seq;
 import io.vavr.collection.Set;
 import io.vavr.control.Option;
 import java.net.URI;
+import java.nio.file.Path;
 import java.util.Collections;
 import java.util.function.BiFunction;
 import org.jetbrains.bsp.bazel.info.BspTargetInfo.TargetInfo;
 import org.jetbrains.bsp.bazel.server.sync.BazelPathsResolver;
 import org.jetbrains.bsp.bazel.server.sync.dependencytree.DependencyTree;
+import org.jetbrains.bsp.bazel.server.sync.languages.JVMLanguagePluginParser;
 import org.jetbrains.bsp.bazel.server.sync.languages.LanguagePlugin;
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaLanguagePlugin;
 import org.jetbrains.bsp.bazel.server.sync.languages.java.JavaModule;
@@ -86,6 +88,11 @@ public class ScalaLanguagePlugin extends LanguagePlugin<ScalaModule> {
 
     buildTarget.setDataKind(BuildTargetDataKind.SCALA);
     buildTarget.setData(scalaBuildTarget);
+  }
+
+  @Override
+  public Option<Path> calculateSourceRoot(Path source) {
+    return JVMLanguagePluginParser.calculateJVMSourceRoot(source, true);
   }
 
   public Option<ScalacOptionsItem> toScalacOptionsItem(Module module) {